requirements.txt
absl-py==0.8.0
antlr4-python3-runtime==4.7.2
appdirs==1.4.3
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==1.5.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
hafnian==0.6.1
html5lib==0.9999999
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lightgbm==2.2.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
mpmath==1.1.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
quantum-blackbird==0.2.1
scikit-learn==0.21.3
scipy==1.2.2
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
StrawberryFields==0.9.0
sympy==1.4
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.6
widgetsnbextension==3.5.1
wrapt==1.11.2
